+++ .eslintrc.js
@@ -1,26 +1,12 @@
 module.exports = {
-  extends: 'erb',
+  extends: 'airbnb',
+  env: {
+    browser: true,
+  },
   rules: {
-    // A temporary hack related to IDE not resolving correct package.json
-    'import/no-extraneous-dependencies': 'off',
-  },
-  parserOptions: {
-    ecmaVersion: 2020,
-    sourceType: 'module',
-    project: './tsconfig.json',
-    tsconfigRootDir: __dirname,
-    createDefaultProgram: true,
-  },
-  settings: {
-    'import/resolver': {
-      // See https://github.com/benmosher/eslint-plugin-import/issues/1396#issuecomment-575727774 for line below
-      node: {},
-      webpack: {
-        config: require.resolve('./.erb/configs/webpack.config.eslint.js'),
-      },
-    },
-    'import/parsers': {
-      '@typescript-eslint/parser': ['.ts', '.tsx'],
-    },
+    'no-console': 'off',
+    'global-require': 'off',
+    'import/no-dynamic-require': 'off',
+    // 'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx'] }],
   },
 };