+++ src/api/preload.js
@@ -4,8 +4,8 @@ const { contextBridge, ipcRenderer } = require('electron');
 contextBridge.exposeInMainWorld('electron', {
   notificationApi: {
     sendNotification(message) {
-      console.log(message)
+      console.log(message);
       ipcRenderer.send('notify', message);
-    }
-  }
+    },
+  },
 });